requirements.txt
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
mutagen==1.47.0
pillow==10.4.0
pycryptodomex==3.20.0
pydub==0.25.1
pytube==15.0.0
requests==2.32.3
tqdm==4.66.5
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.8.6
